"fix" an error when loading in IA v2 interface
Honestly this is a hack. IA is kindly providing us with an object that
has split out the item name, file name, and a few other things. We treat
it as a path (by calling its special toString() method) and then later
split that path into path components to find the item name. Yay.