git rid of 'raw' param to fetch_file
It didn't really do what it sounded like it did; it just converted the
response to an Int8Array. This would only work if the result type was
either null or 'arraybuffer', and we then have to immediately convert
that into a BrowserFS Buffer. Since we can do that directly from an
ArrayBuffer, this is just extra work. Additionally, for rediculuous
reasons a 0-length Int8Array can't be converted into a Buffer, so we
would have a spurious failure if we got a 0-byte response. This then is
the very definition of an anti-feature.