fix scaling at least for MAME
Turns out that a program that uses emscripten's SDL port to create a
"window" (by calling SDL_CreateWindow) gets exactly the canvas size
that they requested, and emscripten actually resets the canvas back to
this size on certain events (such as window resize).

This means that the application must request the scaled size, or
explictly resize the canvas using a different API. Since MAME accepts a
-resolution command-line argument, I can at least make it request a
window of the scaled size. It's not as efficient as letting the
browser scale the window, but it works.