IndexedDB is now used as a cache for game files
Emularity had been storing downloaded game files into the IndexedDB,
but if the file was already present then it would download it again
and overwrite it. Now it will skip the download if the file exists.

At present this only applies to files downloaded using the “saveFile”
method, not the “mountZip” method. The “mountZip” method doesn’t save
the zip file anywhere, so no caching is currently done.

Potentially this could cause you problems if a file gets corrupted or
edited, and reloading the page won’t fix it. Emularity doesn‘t provide
a UI for diagnosing or fixing such problems at present, so you may
want to implement some way to flush the IndexedDB containing the
filesystem.