fix: don't double-load files for MAME items
This prevents us from loading a file as both a driver and a game file,
preventing double-loading .chd files when `meta.emulator_ext` is `chd`.

In these cases, it appears that the files aren't needed as driver files,
so we can safely just load them as game files and ignore the duplicates
we'd get by loading all .chd files as drivers.